+++ packages/react/src/components/Avatar/index.ts
@@ -1,1 +1,2 @@
 export { default as Avatar } from './avatar';
+export * from './Avatar.types';
