+++ packages/react/src/components/Table/ColumnTable.tsx
@@ -2,7 +2,7 @@ import { ColumnTable } from '@cos-ui/primitives';
 import { ComponentProps } from 'react';
 import styled, { css } from 'styled-components';
 
-import Text from '../Text';
+import { Text } from '../Text';
 
 const StyledColumnTableRoot = styled(ColumnTable)`
   width: 100%;
