+++ packages/react/src/components/Table/RowTable.tsx
@@ -2,7 +2,7 @@ import { RowTable } from '@cos-ui/primitives';
 import { ComponentProps } from 'react';
 import styled, { css } from 'styled-components';
 
-import Text from '../Text';
+import { Text } from '../Text';
 
 const StyledRowTableRoot = styled(RowTable)`
   width: 100%;
