+++ packages/react/src/components/Table/columnTable.stories.tsx
@@ -4,7 +4,7 @@ import { ColumnTable } from '.';
 
 import { noticeList } from '../../constants/mock';
 import { FlexBox } from '../FlexBox';
-import Text from '../Text';
+import { Text } from '../Text';
 
 export default {
   title: 'React/Table',
