+++ packages/react/src/components/Text/index.ts
@@ -1,0 +1,9 @@
+import TextRoot from './Text';
+import TextHighlight from './TextHighlight';
+
+const Text = Object.assign(TextRoot, {
+  Highlight: TextHighlight,
+});
+
+export { Text };
+export * from './Text.types';
