+++ packages/react/src/index.ts
@@ -13,7 +13,7 @@ export { default as Paper } from './components/Paper';
 export { default as Radio } from './components/Radio';
 export { BasicTab, BorderTab } from './components/Tab';
 export { ColumnTable, RowTable, BasicTable } from './components/Table';
-export { default as Text } from './components/Text';
+export * from './components/Text';
 export { default as Icon } from './components/Icon';
 export { default as Slider } from './components/Slider';
 export { default as cosStyle, CosStyle } from './styles/GlobalStyle';
