+++ packages/react/src/styles/styles.types.ts
@@ -1,0 +1,5 @@
+import { SizeSX } from './size';
+import { SpacingSX } from './spacing';
+import { Palette } from './theme';
+
+export type { SizeSX, SpacingSX, Palette };
